thor_arquivista_caixa_de_ferramentas/requirements.txt

ttkbootstrap>=1.10.1
pymongo>=4.10.1
tqdm>=4.66.4
PyYAML>=6.0.2
python-dotenv>=1.0.1
Pillow
